+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 - Method `count_parameters` to `TModule`.
 - Option `padding_idx` to `indices_to_onehot` function.
 - Function `dict_list_to_list_dict` to collections utils.
+- Class `PickleDataset` and function `pack_to_pickle` to utils.
 
 ### Fixed
 - Remove invalid test file.